requirements.txt
pymongo
python-dotenv
tabulate